requirements.txt
redis
fakeredis
pandas
pathvalidate
vtracer
python-dotenv
pytest